+++ actionpack/lib/action_dispatch/railtie.rb
@@ -25,6 +25,8 @@ module ActionDispatch
       'X-Content-Type-Options' => 'nosniff'
     }
 
+    config.eager_load_namespaces << ActionDispatch
+
     initializer "action_dispatch.configure" do |app|
       ActionDispatch::Http::URL.tld_length = app.config.action_dispatch.tld_length
       ActionDispatch::Request.ignore_accept_header = app.config.action_dispatch.ignore_accept_header