+++ actionpack/lib/action_view/railtie.rb
@@ -9,6 +9,8 @@ module ActionView
     config.action_view.javascript_expansions = { :defaults => %w(jquery jquery_ujs) }
     config.action_view.embed_authenticity_token_in_remote_forms = false
 
+    config.eager_load_namespaces << ActionView
+
     initializer "action_view.embed_authenticity_token_in_remote_forms" do |app|
       ActiveSupport.on_load(:action_view) do
         ActionView::Helpers::FormTagHelper.embed_authenticity_token_in_remote_forms =