+++ railties/test/application/rake_test.rb
@@ -55,6 +55,29 @@ module ApplicationTests
       assert_match "Doing something...", output
     end
 
+    def test_does_not_explode_when_accessing_a_model_with_eager_load
+      add_to_config <<-RUBY
+        config.eager_load = true
+
+        rake_tasks do
+          task :do_nothing => :environment do
+            Hello.new.world
+          end
+        end
+      RUBY
+
+      app_file "app/models/hello.rb", <<-RUBY
+      class Hello
+        def world
+          puts "Hello world"
+        end
+      end
+      RUBY
+
+      output = Dir.chdir(app_path){ `rake do_nothing` }
+      assert_match "Hello world", output
+    end
+
     def test_code_statistics_sanity
       assert_match "Code LOC: 5     Test LOC: 0     Code to Test Ratio: 1:0.0",
         Dir.chdir(app_path){ `rake stats` }