+++ railties/CHANGELOG.md
@@ -5,6 +5,8 @@
     default_url_options[:script_name] to set proper application's mount point by
     yourself. *Piotr Sarnacki*
 
+*   `config.threadsafe!` is deprecated in favor of `config.eager_load` which provides a more fine grained control on what is eager loaded *José Valim*
+
 *   The migration generator will now produce AddXXXToYYY/RemoveXXXFromYYY migrations with references statements, for instance
 
         rails g migration AddReferencesToProducts user:references supplier:references{polymorphic}